pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.8.0"
asyncio = "^3.4.3"
python-dotenv = "^1.0.1"
motor = "^3.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pylint = "^3.0.2"
isort = "^5.13.0"
black = "^23.11.0"
model-bakery = "^1.17.0"
pylint-pytest = "^1.1.7"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79
lines_after_imports = 2

[tool.pylint]
recursive="y"
ignore = [".venv"]
max-line-length = 79
disable = ["C0114", "C0115", "C0116", "W0511"]
